pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnirt"
version = "0.1.0a0"
description = "Open image & video generation runtime with CUDA and Ascend backends."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "datascale-ai" }
]
dependencies = [
  "imageio>=2.34",
  "imageio-ffmpeg>=0.4.9",
  "numpy>=1.24",
  "PyYAML>=6.0",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
runtime = [
  "diffusers>=0.27",
  "huggingface_hub>=0.20",
  "safetensors>=0.4",
  "torch>=2.1",
  "transformers>=4.36",
]

[project.scripts]
omnirt = "omnirt.cli.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
